Refactor clients endpoint in main.py
- Replaced the existing get_clients function with clients_endpoint to streamline client retrieval.
- Improved code organization by importing get_clients from the appropriate module.
